Customisations and Enhancements (#320)
* fix: change parcel entry point *.html
* remove extra checks for config.app
* new flag `--[no-]serve    start frontend server`
* add default values to flags
* explicitly set app config test values

+++ src/lib/runDev.js
@@ -120,6 +120,10 @@ async function runDev (args = [], config, options = {}, log = () => {}) {
   const withBackend = config.app.hasBackend && !skipActions
   const isLocal = !config.actions.devRemote // applies only for backend
 
+  aioLogger.debug(`hasFrontend ${hasFrontend}`)
+  aioLogger.debug(`withBackend ${withBackend}`)
+  aioLogger.debug(`isLocal ${isLocal}`)
+
   // port for UI
   const uiPort = parseInt(args[0]) || parseInt(process.env.PORT) || 9080
 
@@ -179,37 +183,38 @@ async function runDev (args = [], config, options = {}, log = () => {}) {
       }
       await utils.writeConfig(devConfig.web.injectedConfig, urls)
 
-      log('starting local frontend server ..')
-      const entryFile = path.join(devConfig.web.src, 'index.html')
+      if (!options.skipServe) {
+        log('starting local frontend server ..')
+        const entryFile = path.join(devConfig.web.src, '*.html')
 
-      // our defaults here can be overridden by the bundleOptions passed in
-      // bundleOptions.https are also passed to bundler.serve
-      const parcelBundleOptions = {
-        cache: false,
-        outDir: devConfig.web.distDev,
-        contentHash: false,
-        watch: true,
-        minify: false,
-        logLevel: 1,
-        ...bundleOptions
+        // our defaults here can be overridden by the bundleOptions passed in
+        // bundleOptions.https are also passed to bundler.serve
+        const parcelBundleOptions = {
+          cache: false,
+          outDir: devConfig.web.distDev,
+          contentHash: false,
+          watch: true,
+          minify: false,
+          logLevel: 1,
+          ...bundleOptions
+        }
+
+        let actualPort = uiPort
+        const uiBundler = new Bundler(entryFile, parcelBundleOptions)
+        cleanup.add(() => uiBundler.stop(), 'stopping parcel watcher...')
+        const uiServer = await uiBundler.serve(uiPort, bundleOptions.https)
+        actualPort = uiServer.address().port
+        const uiServerTerminator = httpTerminator.createHttpTerminator({
+          server: uiServer
+        })
+        cleanup.add(() => uiServerTerminator.terminate(), 'stopping ui server...')
+
+        if (actualPort !== uiPort) {
+          log(`Could not use port:${uiPort}, using port:${actualPort} instead`)
+        }
+        frontEndUrl = `${bundleOptions.https ? 'https:' : 'http:'}//localhost:${actualPort}`
+        log(`local frontend server running at ${frontEndUrl}`)
       }
-
-      let actualPort = uiPort
-      const uiBundler = new Bundler(entryFile, parcelBundleOptions)
-      cleanup.add(() => uiBundler.stop(), 'stopping parcel watcher...')
-      const uiServer = await uiBundler.serve(uiPort, bundleOptions.https)
-      actualPort = uiServer.address().port
-      const uiServerTerminator = httpTerminator.createHttpTerminator({
-        server: uiServer
-      })
-      cleanup.add(() => uiServerTerminator.terminate(), 'stopping ui server...')
-
-      if (actualPort !== uiPort) {
-        log(`Could not use port:${uiPort}, using port:${actualPort} instead`)
-      }
-
-      frontEndUrl = `${bundleOptions.https ? 'https:' : 'http:'}//localhost:${actualPort}`
-      log(`local frontend server running at ${frontEndUrl}`)
     }
 
     log('setting up vscode debug configuration files...')
@@ -218,7 +223,7 @@ async function runDev (args = [], config, options = {}, log = () => {}) {
     cleanup.add(() => vscodeConfig.cleanup(), 'cleaning up vscode debug configuration files...')
 
     if (!isLocal && !hasFrontend) {
-      // not local + ow is not running => need to explicitely wait for CTRL+C
+      // not local + ow is not running => need to explicitly wait for CTRL+C
       // trick to avoid termination
       const dummyProc = execa('node')
       cleanup.add(() => dummyProc.kill(), 'stopping sigint waiter...')
